billing

+++ app/src/main/java/com/it10x/foodappgstav7_06/ui/payment/PaymentInput.kt
@@ -1,8 +1,8 @@
 package com.it10x.foodappgstav7_06.ui.payment
 
 data class PaymentInput(
-    val mode: String,       // CASH | UPI | CARD | CREDIT
-    val amount: Double
+    val mode: String,
+    val amount: Long   // ✅ ALWAYS PAISE
 )
